refactor AdvancedSummary into its own file
